fix libbinder conflict system lib. Rename libbinder to libbinder_vendor
in this change we are correcting the conflict libbinder.so on Erfangsi ports. In a good way, you need to figure out why libbinder in the libqdutils and libqservice libraries does not load from apex. Rename libbinder to libbinder_vendor and build its vendor copy. We also remove unnecessary library headers and move them and connect them locally.